Update README.md
Ever seen an ant scurry by and wonder what it was doing, where it was going or where it came from? This application will analyze the chances that you will notice the nature of such things in any given minute, hour, day, week, year or even an entire lifetime. It will take into account location, current weather, weather forecasts and also pull from historical data the odds that someone noticed an event small enough that only they as the user noticed. Many of these things we never notice. This application will lessen the odds that the user is likely to miss a very small phenomenon, thus causing the user to be aware that something VERY SMALL is about to happen in their "user experience" of life. It will be able to be "dialed" or scaled in or out as to the discretion of the end user.